Safari-609.1.11.1/LayoutTests/js/no-iterator-constructors-expected.txt

This test makes sure we aren't putting the iterator constructors on the global object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'ArrayIterator' in this is false
PASS 'ArgumentsIterator' in this is false
PASS 'MapIterator' in this is false
PASS 'SetIterator' in this is false
PASS successfullyParsed is true

TEST COMPLETE
